refactor all poll creation screens to separate files and export common style sheet elements
